Move python wrapper to contrib/python-bindings.
