+++ contrib/python-bindings/source/cell_accessor_wrapper.cc
@@ -13,10 +13,10 @@
 //
 // ---------------------------------------------------------------------
 
+#include <point_wrapper.h>
+#include <triangulation_wrapper.h>
+#include <cell_accessor_wrapper.h>
 #include <boost/python.hpp>
-#include <deal.II/python/point_wrapper.h>
-#include <deal.II/python/triangulation_wrapper.h>
-#include <deal.II/python/cell_accessor_wrapper.h>
 
 DEAL_II_NAMESPACE_OPEN
