+++ contrib/python-bindings/source/export_point.cc
@@ -13,10 +13,10 @@
 //
 // ---------------------------------------------------------------------
 
-#include <boost/python.hpp>
-#include <deal.II/python/point_wrapper.h>
+#include <point_wrapper.h>
 #include <deal.II/base/point.h>
 #include <deal.II/base/exceptions.h>
+#include <boost/python.hpp>
 
 DEAL_II_NAMESPACE_OPEN
