+++ contrib/python-bindings/tests/point_wrapper.py
