+++ source/CMakeLists.txt
@@ -48,9 +48,6 @@ ADD_SUBDIRECTORY(integrators)
 ADD_SUBDIRECTORY(matrix_free)
 ADD_SUBDIRECTORY(meshworker)
 ADD_SUBDIRECTORY(opencascade)
-IF(DEAL_II_WITH_PYTHON)
-  ADD_SUBDIRECTORY(python)
-ENDIF()
 
 FOREACH(build ${DEAL_II_BUILD_TYPES})
   STRING(TOLOWER ${build} build_lowercase)
